requirements.txt
fastapi
uvicorn
python-dotenv
twilio
websockets
pyngrok
pydub
jinja2
aiofiles
google-cloud-speech

